perm-checkify

+++ src/BaseClient/ClientHelperModules/requestHandler/channels/deleteAllReactions.ts
@@ -3,15 +3,40 @@ import error from '../../error.js';
 import { API } from '../../../Client.js';
 import cache from '../../cache.js';
 
+import getBotMemberFromGuild from '../../getBotMemberFromGuild.js';
+import requestHandlerError from '../../requestHandlerError.js';
+
 /**
  * Deletes all reactions from a message in a channel.
  * @param message The message to delete reactions from.
  * @returns A promise that resolves with the deleted message, or rejects with a DiscordAPIError.
  */
-export default async (message: Discord.Message<true>) =>
- (cache.apis.get(message.guild.id) ?? API).channels
+export default async (message: Discord.Message<true>) => {
+ if (!canDeleteAllReactions(message.channel, await getBotMemberFromGuild(message.guild))) {
+  const e = requestHandlerError(
+   `Cannot delete all reactions of messages in ${message.guild.name} / ${message.guild.id}`,
+   [Discord.PermissionFlagsBits.ManageMessages],
+  );
+
+  error(message.guild, e);
+  return e;
+ }
+
+ return (cache.apis.get(message.guild.id) ?? API).channels
   .deleteAllMessageReactions(message.channel.id, message.id)
   .catch((e) => {
    error(message.guild, new Error((e as Discord.DiscordAPIError).message));
    return e as Discord.DiscordAPIError;
   });
+};
+
+/**
+ * Checks if the user has the permission to delete all reactions in a channel.
+ * @param channel - The guild-based channel to check.
+ * @param me - The guild member representing the user.
+ * @returns A boolean indicating whether the user has the permission to delete all reactions.
+ */
+export const canDeleteAllReactions = (
+ channel: Discord.GuildBasedChannel,
+ me: Discord.GuildMember,
+) => me.permissionsIn(channel).has(Discord.PermissionFlagsBits.ManageMessages);
